+++ changelog.md
@@ -53,7 +53,7 @@ Also EnableSolanaAddressLookupTable feature flag should be set.
 * [4254](https://github.com/zeta-chain/node/pull/4254) - add additional support for zetaclient public DNS name
 * [4342](https://github.com/zeta-chain/node/pull/4342) - add metrics for monitoring inbound voting through blockscan and trackers
 * [4384](https://github.com/zeta-chain/node/pull/4384) - disable TSS service in dry mode
-
+* [4359](https://github.com/zeta-chain/node/pull/4359) - add chaos mode for ZetaClient
 
 ### Tests